+++ docs/demo/FR/index.jsx
@@ -1,6 +1,9 @@
 import React, { useState } from 'react';
+// 使用 Ant Design 风格
 import FormRender from 'form-render/lib/antd';
+// 使用 Fusion 风格
 // import FormRender from 'form-render/lib/fusion';
+// import '@alifd/next/dist/next.min.css';
 
 const Demo = ({ schema = {} }) => {
   const [formData, setFormData] = useState(() => schema.formData);