+++ src/widgets/antd/upload.jsx
@@ -2,36 +2,41 @@ import React from 'react';
 import { UploadOutlined } from '@ant-design/icons';
 import { Upload, message, Button } from 'antd';
 
-export default function input(p) {
+export default function upload({
+  action,
+  value,
+  name,
+  onChange,
+  options = {},
+}) {
+  const _action = action || (options && options.action);
+  const _className = `fr-upload-file ${options ? options.className : ''}`;
   const props = {
     name: 'file',
-    action: p.action,
-    enctype: 'multipart/form-data',
-    withCredentials: true,
-    type: 'file',
-    onChange(info) {
+    action: _action,
+    className: _className,
+    onChange: info => {
       if (info.file.status === 'done') {
         message.success(`${info.file.name} 上传成功`);
-        p.onChange(p.name, info.file.response.url);
+        onChange(name, info.file.response.url);
       } else if (info.file.status === 'error') {
         message.error(`${info.file.name} 上传失败`);
       }
     },
     onRemove() {
-      p.onChange(p.name, '');
+      onChange(name, '');
     },
+    ...options,
   };
 
   return (
     <div className="fr-upload-mod">
-      <Upload {...props} className="fr-upload-file">
-        <Button>
-          <UploadOutlined /> 上传
-        </Button>
+      <Upload {...props}>
+        <Button icon={<UploadOutlined />}>上传</Button>
       </Upload>
-      {p.value && (
+      {value && (
         <a
-          href={p.value}
+          href={value}
           target="_blank"
           rel="noopener noreferrer"
           className="fr-upload-preview"